refactor: split tui/app.rs from 1648 to 1002 lines
Extract three focused modules from the monolithic app.rs:
- settings_state.rs (309 lines): SettingField enum, SettingsState struct
- sparkline_data.rs (257 lines): all sparkline builders and helpers
- rollup.rs: aggregate_summaries_to_models, merge_model_usages
- types.rs: GroupBy enum (shared domain concept)
