+++ src/tui/widgets/usage_table.rs
@@ -127,17 +127,17 @@ pub fn render(frame: &mut Frame, area: Rect, app: &App) {
             // Use raw_model for API provider inference (retains routing
             // prefix like "vertexai."), normalized model for display name.
             let (name_col, api_col, client_col) = match app.group_by {
-                crate::tui::app::GroupBy::Model => (
+                crate::types::GroupBy::Model => (
                     display::display_model(&mu.model),
                     display::infer_api_provider(mu.effective_raw_model()).to_string(),
                     String::new(),
                 ),
-                crate::tui::app::GroupBy::ModelClient => (
+                crate::types::GroupBy::ModelClient => (
                     display::display_model(&mu.model),
                     display::infer_api_provider(mu.effective_raw_model()).to_string(),
                     display::display_client(&mu.provider).into_owned(),
                 ),
-                crate::tui::app::GroupBy::Client => (
+                crate::types::GroupBy::Client => (
                     display::display_client(&mu.provider).into_owned(),
                     String::new(),
                     String::new(),
